fix: match suggested prompts to source language
